pyproject.toml
[project]
name = "django-for-vue-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.3"
dependencies = [
    "django>=5.2.4",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-yasg>=1.21.10",
    "django-filter>=24.1",
]

[dependency-groups]
dev = [
    "django-stubs>=5.2.1",
    "pytest-django>=4.11.1",
]
